+++ packages/vite-plugin-cloudflare/src/plugins/config.ts
@@ -10,6 +10,7 @@ import {
 import { normalizePath } from "vite";
 import { hasAssetsConfigChanged } from "../asset-config";
 import { createBuildApp, removeAssetsField } from "../build";
+import { buildOutputContainers } from "../build-output-containers";
 import {
 	cloudflareBuiltInModules,
 	createCloudflareEnvironmentOptions,
@@ -271,6 +272,11 @@ export const configPlugin = createPlugin("config", (ctx) => {
 						);
 					}
 				}
+
+				await buildOutputContainers(
+					ctx.resolvedPluginConfig,
+					builder.config.root
+				);
 			},
 		},
 	};